requirements.txt
# Project requirements. Install by command: pip install -r requirements.txt
aiogram==3.26.0
python-dotenv==1.2.2
aiosqlite==0.22.1
ollama==0.6.1